requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
firebase-admin==6.3.0
python-dotenv==1.0.0
stripe==7.11.0
